# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TransitOps fleet operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List vehicles
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/VehicleStatus"
        - name: region
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Register a new vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleUpdate"
      responses:
        "200":
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      summary: Delete a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    VehicleStatus:
      type: string
      enum: [Available, On Trip, In Shop, Retired]
    Vehicle:
      type: object
      required:
        [
          id,
          registrationNumber,
          name,
          type,
          maxLoadCapacity,
          odometer,
          acquisitionCost,
          status,
          region,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: integer }
        registrationNumber: { type: string }
        name: { type: string }
        type: { type: string }
        maxLoadCapacity: { type: number }
        odometer: { type: number }
        acquisitionCost: { type: number }
        status:
          $ref: "#/components/schemas/VehicleStatus"
        region: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    VehicleInput:
      type: object
      required:
        [
          registrationNumber,
          name,
          type,
          maxLoadCapacity,
          acquisitionCost,
          region,
        ]
      properties:
        registrationNumber: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        type: { type: string, minLength: 1 }
        maxLoadCapacity: { type: number, minimum: 0 }
        odometer: { type: number, minimum: 0 }
        acquisitionCost: { type: number, minimum: 0 }
        region: { type: string, minLength: 1 }
    VehicleUpdate:
      type: object
      properties:
        registrationNumber: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        type: { type: string, minLength: 1 }
        maxLoadCapacity: { type: number, minimum: 0 }
        odometer: { type: number, minimum: 0 }
        acquisitionCost: { type: number, minimum: 0 }
        region: { type: string, minLength: 1 }
        status:
          $ref: "#/components/schemas/VehicleStatus"
...