# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TransitOps fleet operations API
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleStatus } from './vehicleStatus';

export interface Vehicle {
  id: number;
  registrationNumber: string;
  name: string;
  type: string;
  maxLoadCapacity: number;
  odometer: number;
  acquisitionCost: number;
  status: VehicleStatus;
  region: string;
  createdAt: Date;
  updatedAt: Date;
}