# lib/api-zod/src/generated/types/vehicleUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * TransitOps fleet operations API
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleStatus } from './vehicleStatus';

export interface VehicleUpdate {
  /** @minLength 1 */
  registrationNumber?: string;
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  type?: string;
  /** @minimum 0 */
  maxLoadCapacity?: number;
  /** @minimum 0 */
  odometer?: number;
  /** @minimum 0 */
  acquisitionCost?: number;
  /** @minLength 1 */
  region?: string;
  status?: VehicleStatus;
}